+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Split configuration by moving various selections into
   `$XDG_CACHE_HOME/notnow/ui-state.json`
+- Warn if unsaved changes are present when trying to quit program
 
 
 0.3.2